Move converter class to converters file.
