+++ src/DynamoCoreWpf/Views/Preview/Watch3DView.xaml.cs
@@ -1061,28 +1061,4 @@ namespace Dynamo.Controls
         public MeshGeometry3D SelectedMesh { get; set; }
         public BillboardText3D Text { get; set; }
     }
-
-    internal class Watch3DBackgroundColorConverter : IValueConverter
-    {
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            var homeColor = (System.Windows.Media.Color) SharedDictionaryManager.DynamoColorsAndBrushesDictionary["WorkspaceBackgroundHome"];
-            var customColor = (System.Windows.Media.Color)SharedDictionaryManager.DynamoColorsAndBrushesDictionary["WorkspaceBackgroundCustom"];
-
-            //parameter will contain a true or false
-            //whether this is the home space
-            if ((bool) value)
-            {
-                return homeColor.ToColor4() ;
-            }
-                
-            return customColor.ToColor4();
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter,
-          CultureInfo culture)
-        {
-            return null;
-        }
-    }
 }
